pyproject.toml
[project]
name = "webots-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi",
	"uvicorn",
	"streamlit",
	"requests",
	"pydantic",
	"pyyaml",
	"google-genai",
]
